+++ demos/employees/src/main/java/is/codion/framework/demos/employees/testing/scenarios/UpdateEmployee.java
@@ -23,7 +23,6 @@ import is.codion.framework.demos.employees.domain.Employees.Department;
 import is.codion.framework.demos.employees.domain.Employees.Employee;
 import is.codion.framework.demos.employees.model.EmployeesAppModel;
 import is.codion.framework.domain.entity.Entity;
-import is.codion.framework.domain.entity.exception.ValidationException;
 import is.codion.framework.domain.test.DefaultEntityFactory;
 import is.codion.framework.domain.test.DomainTest.EntityFactory;
 import is.codion.swing.framework.model.SwingEntityModel;
@@ -36,7 +35,7 @@ public final class UpdateEmployee extends AbstractPerformer {
 	private final Random random = new Random();
 
 	@Override
-	public void perform(EmployeesAppModel application) throws ValidationException {
+	public void perform(EmployeesAppModel application) {
 		SwingEntityModel departmentModel = application.entityModel(Department.TYPE);
 		selectRandomRow(departmentModel.tableModel());
 		SwingEntityModel employeeModel = departmentModel.detailModel(Employee.TYPE);