+++ demos/world/src/main/java/is/codion/framework/demos/world/model/CityTableModel.java
@@ -25,7 +25,6 @@ import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.demos.world.domain.api.World.City;
 import is.codion.framework.demos.world.domain.api.World.Country;
 import is.codion.framework.domain.entity.Entity;
-import is.codion.framework.domain.entity.exception.ValidationException;
 import is.codion.swing.common.model.worker.ProgressWorker.ProgressReporter;
 import is.codion.swing.common.model.worker.ProgressWorker.ProgressResultTask;
 import is.codion.swing.framework.model.SwingEntityTableModel;
@@ -105,7 +104,7 @@ public final class CityTableModel extends SwingEntityTableModel {
 		}
 
 		@Override
-		public Void execute(ProgressReporter<String> progressReporter) throws ValidationException, IOException {
+		public Void execute(ProgressReporter<String> progressReporter) throws IOException {
 			Collection<Entity> updatedCities = new ArrayList<>();
 			CityEditModel editModel = editModel();
 			Iterator<Entity> citiesIterator = cities.iterator();