+++ framework/domain/src/main/java/is/codion/framework/domain/entity/exception/ValidationException.java
@@ -25,7 +25,7 @@ import static java.util.Objects.requireNonNull;
 /**
  * An exception used to indicate that an attribute value is invalid.
  */
-public class ValidationException extends Exception {
+public class ValidationException extends RuntimeException {
 
 	private final Attribute<?> attribute;
 	private final Object value;